# www/docs/api.yml
openapi: 3.0.*
info:
  title: Representation API
  version: 0.0.1
servers:
  - url: https://mira.isi.edu
paths:
...
  /datasets/{dataset_name}/resources:
    post:
      summary: create resource
      tags: [resource]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/path/dataset_name"
      requestBody:
        required: true
        content:
          "multipart/form-data":
            schema:
              type: object
              properties:
                resource_file:
                  type: string
                  format: binary
                resource_id:
                  type: string
                  resource_type:
                    $ref: "#/components/schemas/Resource/properties/type"
                  extra:
                    type: object
      responses:
        401:
          description: Unauthorized request
        200:
          content:
            'application/json':
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: ["success"]
                  resource:
                    $ref: "#/components/schemas/Resource"
  /datasets/{dataset_name}/resources/{resource_id}:
    delete:
      summary: delete a resource
      tags: [resource]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/path/dataset_name"
        - $ref: "#/components/parameters/path/resource_id"
      responses:
        401:
          description: Unauthorized request
    get:
      summary: get resource information
      tags: [resource]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/path/dataset_name"
        - $ref: "#/components/parameters/path/resource_id"
      responses:
        401:
          description: Unauthorized request
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
...
  schemas:
...
    Resource:
      type: object
      properties:
        id:
          type: string
          description: id of resource
        type:
          type: string
          enum: ["csv", "netcdf4", "json"]
          description: type of resource
...